✨ feat(__init__.py): add user blueprint to the app 🔨 refactor(routes.py): remove unused imports and endpoints 🚚 chore(navigator.html): update urls to use user blueprint instead of main blueprint 🆕 feat(user/__init__.py): create user blueprint and import routes
✨ feat(routes.py): add user, wishlist, wishlist_product, cart, cart_quantity, and checkout routes for user module. Add CartSchema and EmptySchema for response. Add before_app_request decorator to set search form to global config.
