+++ app/__init__.py
@@ -54,6 +54,8 @@ def create_app(config_name:str = 'default'):
     app.register_blueprint(admin_bp)
     from app.auth import bp as auth_bp
     app.register_blueprint(auth_bp)
+    from app.user import bp as user_bp
+    app.register_blueprint(user_bp)
     from app.api import bp as api_bp
     app.register_blueprint(api_bp, url_prefix='/api')
