+++ app/main/forms.py
@@ -3,6 +3,7 @@ from flask_wtf import FlaskForm
 from wtforms import StringField
 from wtforms.validators import DataRequired
 
+
 class SearchForm(FlaskForm):
     """Search form.
     """
