+++ app/user/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('user', __name__)
+
+from app.user import routes
